Correctly avoid empty tiles + Handle bigger tiles without crashing

+++ src/IO/Raster.h
@@ -55,6 +55,6 @@ private:
     uint width;
     uint height;
     float pixelSize;
-    float noDataValue;
+    double noDataValue;
     double geoTransform[6];
 };
